First commit
Run skribbler.py [keyword] to create a window displaying 1 of 10 images
that matched the keyword. Use the left and right arrow keys to navigate
between them and return to select an image. In skribbl.io, click the
white square in the palette followed by the top-left then bottom-right
corners of the canvas.